fix(bench): correct _run_one_spec qrel_index type annotation
The FIX 2 refactor extracted _run_one_spec from run_bench but typed
qrel_index as dict[str, dict[str, int]]; QrelIndex is actually
dict[str, dict[str, float]] (BEIR uses graded relevance, the metrics
take floats). Mypy's dict invariance flagged the mismatch when
recall_at_k / ndcg_at_k were called inside the helper.

+++ src/vdbbench/bench/runner.py
@@ -336,7 +336,7 @@ def _run_one_spec(
     encoded: EncodedBundle,
     pids: list[str],
     qids: list[str],
-    qrel_index: dict[str, dict[str, int]],
+    qrel_index: dict[str, dict[str, float]],
     timing_rows: list[QueryTiming],
     summary_rows: list[RunSummary],
 ) -> None:
